+++ app/Http/Controllers/Admin/CarsController.php
@@ -10,6 +10,7 @@ use App\Contracts\Services\FlashMessageContract;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CarRequest;
 use App\Http\Requests\TagsRequest;
+use App\Models\Car;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
@@ -24,6 +25,8 @@ class CarsController extends Controller
 
     public function index(): Factory|View|Application
     {
+        $this->authorize('viewAny', Car::class);
+
         $cars = $this->repository->findAll()->sortByDesc('updated_at');
         
         return view('pages.admin.cars.list', ['cars' => $cars]);
@@ -31,6 +34,8 @@ class CarsController extends Controller
 
     public function create(): Factory|View|Application
     {
+        $this->authorize('create', Car::class);
+
         return view('pages.admin.cars.create');
     }
 
@@ -40,6 +45,8 @@ class CarsController extends Controller
         FlashMessageContract $flashMessage,
         CarCreationServiceContract $createServise,
     ): RedirectResponse {
+        $this->authorize('create', Car::class);
+
         $fields = $request->validated();
         $categories = $request->get('categories');
         $tags = $tagsRequest->get('tags', []);
@@ -51,14 +58,11 @@ class CarsController extends Controller
         return redirect()->route('admin.cars.index');
     }
 
-    public function show(int $id)
-    {
-        //
-    }
-
     public function edit(int $id): Factory|View|Application
     {
         $car = $this->repository->getById($id, ['categories', 'image', 'images', 'tags']);
+
+        $this->authorize('update', $car);
 
         return view('pages.admin.cars.edit', ['car' => $car]);
     }
@@ -70,6 +74,8 @@ class CarsController extends Controller
         FlashMessageContract $flashMessage,
         CarUpdateServiceContract $updateServise,
     ): RedirectResponse {
+        $this->authorize('update', [Car::class, $id]);
+
         $fields = $request->validated();
         $categories = $request->get('categories');
         $tags = $tagsRequest->get('tags', []);
@@ -86,6 +92,8 @@ class CarsController extends Controller
         CarRemoveServiceContract $removeService,
         FlashMessageContract $flashMessage,
     ): RedirectResponse {
+        $this->authorize('delete', [Car::class, $id]);
+
         $removeService->delete($id);
 
         $flashMessage->success('Модель удалена');
